orchestrator/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.14.0
python-socketio[asyncio_client]>=5.3.0
pymongo>=3.11.0
pydantic>=1.8.2
docker>=5.0.0
python-dotenv>=0.19.0
httpx>=0.19.0